requirements/base.txt


# Bleeding edge Django
django==1.9.8

# Configuration
django-environ==0.4.0
whitenoise==3.2


# Forms
django-braces==1.9.0
django-crispy-forms==1.6.1
django-floppyforms==1.6.2

# Models
django-model-utils==2.5.1

# Images
Pillow==3.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.26.1


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0


celery==3.1.23
django-mptt==0.8.7
django_compressor==2.0
django-import-export==0.5.1


# Webpack
django-webpack-loader==0.3.3

# Django Suit
django-suit==0.2.25
django-bootstrap3==8.2.3
bootstrap3-datetime==2.4
django-prettyjson==0.3.0

django-autocomplete-light==3.2.1
# Your custom requirements go here

djangorestframework==3.6.3
django-rest-swagger==2.1.2
django-tables2==1.10.0
django-filter==0.13.0
